Time: 10 ms (57.93%) | Memory: 17.4 MB (25.97%) - LeetSync

+++ maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -1,17 +1,17 @@
 class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
-       queue<pair<TreeNode*,long long int>> que;
+       queue<pair<TreeNode*,long int>> que;
        if (root->left!=NULL){que.push({root->left,1});}
        if (root->right!=NULL){que.push({root->right,2});}
-       long long int m = 1;
+       long int m = 1;
        while(que.size())
        {
             int n = que.size();
             TreeNode *temp;
-            long long int rev;
+            long int rev;
             m = max(que.back().second - que.front().second + 1,m);
-            long long int x = que.front().second;
+            long int x = que.front().second;
             for(int i=0;i<n;i++)
             {
                 temp = que.front().first;
